Also consolidate the files
