+++ js/panel-config/automation.js
@@ -1,8 +1,8 @@
 import { h, Component } from 'preact';
 
 import Trigger from './trigger/index.js';
-import Condition from '../common/component/condition/index.js';
-import Script from '../common/component/script/index.js';
+import Condition from './condition/index.js';
+import Script from './script/index.js';
 
 export default class Automation extends Component {
   constructor() {
