+++ js/panel-config/condition/condition_edit.js
