+++ js/panel-config/condition/numeric_state.js
@@ -1,6 +1,6 @@
 import { h, Component } from 'preact';
 
-import { onChangeEvent } from '../../util/event.js';
+import { onChangeEvent } from '../../common/util/event.js';
 
 export default class NumericStateCondition extends Component {
   constructor() {
