+++ js/panel-config/panel-config.js
