+++ js/panel-config/script/action_edit.js
