+++ js/panel-config/script/action_row.js
