+++ js/panel-config/script/index.js
