+++ js/panel-config/trigger/homeassistant.js
