+++ js/panel-config/trigger/index.js
