+++ js/panel-config/trigger/mqtt.js
