+++ js/panel-config/trigger/numeric_state.js
