+++ js/panel-config/trigger/state.js
