+++ js/panel-config/trigger/sun.js
