+++ js/panel-config/trigger/template.js
