+++ js/panel-config/trigger/time.js
