+++ js/panel-config/trigger/trigger_edit.js
