+++ js/panel-config/trigger/trigger_row.js
