+++ js/panel-config/trigger/zone.js
