+++ panels/config/automation/ha-automation-editor.html
@@ -22,8 +22,7 @@
 <link rel='import' href='../../../src/util/hass-mixins.html'>
 
 <link rel="import" href="../ha-config-section.html">
-
-<script src='../../../build-temp/automation-editor.js'></script>
+<link rel="import" href="../ha-config-js.html">
 
 <dom-module id="ha-automation-editor">
   <template>
