+++ panels/config/ha-config-js.html
@@ -1,0 +1,1 @@
+<script src='../../build-temp/panel-config.js'></script>
