+++ panels/config/script/ha-script-editor.html
@@ -22,7 +22,7 @@
 
 <link rel="import" href="../ha-config-section.html">
 
-<script src='../../../build-temp/automation-editor.js'></script>
+<link rel="import" href="../ha-config-js.html">
 
 <dom-module id="ha-script-editor">
   <template>
